Fiexd Eslint for AuthContexed and added useLocation to triger rerender
