+++ src/main.tsx
@@ -4,7 +4,8 @@ import './index.css';
 import App from './App.tsx';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { BrowserRouter } from 'react-router-dom';
-import { AuthProvider } from '@/context/AuthContext.tsx';
+import { AuthProvider } from '@/context/Auth/AuthProvider.tsx';
+
 
 const queryClinet = new QueryClient();
 createRoot(document.getElementById('root')!).render(
